[SPOT-288][REFACTOR] StudyInfoResponseDTO 통합
